requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
CacheControl==0.14.3
cachetools==5.5.2
celery==5.5.3
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==45.0.5
daphne==4.2.1
dj-database-url==2.3.0
Django==5.1.5
django-celery-beat==2.8.1
django-cors-headers==4.6.0
django-redis==6.0.0
django-redis-cache==3.0.1
django-rest-knox==5.0.2
django-rest-passwordreset==1.5.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dynaconf==3.2.7
firebase_admin==7.0.0
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
gunicorn==23.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hvac==2.3.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.4
loguru==0.7.3
msgpack==1.1.1
multidict==6.6.3
packaging==24.2
pillow==11.3.0
prompt_toolkit==3.0.51
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==3.5.3
requests==2.32.3
rsa==4.9.1
service-identity==24.2.0
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
validators==0.34.0
vine==5.1.0
wcwidth==0.2.13
win32_setctime==1.2.0
yarl==1.20.1
zope.interface==7.2
